Separate layers for modularization framework + data
