--- app/src/main/kotlin/com/architects/pokearch/core/data/local/database/entities/PokemonEntity.kt
@@ -1,12 +1,0 @@
-package com.architects.pokearch.core.data.local.database.entities
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class PokemonEntity (
-    @PrimaryKey
-    val id: Int,
-    val name: String,
-    val url: String,
-)
