--- app/src/main/kotlin/com/architects/pokearch/core/data/local/database/entities/converters/StatsHolder.kt
@@ -1,12 +1,0 @@
-package com.architects.pokearch.core.data.local.database.entities.converters
-
-data class StatsHolder (val stats: List<StatsEntity>)
-
-data class StatsEntity(
-    val value: Int,
-    val stat: StatEntity
-)
-
-data class StatEntity(
-    val name: String
-)
