--- app/src/main/kotlin/com/architects/pokearch/core/data/local/database/entities/converters/StatsHolderConverter.kt
@@ -1,11 +1,0 @@
-package com.architects.pokearch.core.data.local.database.entities.converters
-
-import androidx.room.TypeConverter
-import com.google.gson.Gson
-
-class StatsHolderConverter {
-    @TypeConverter
-    fun fromType(stats: StatsHolder): String = Gson().toJson(stats)
-    @TypeConverter
-    fun toType(stats: String): StatsHolder = Gson().fromJson(stats, StatsHolder::class.java)
-}
