--- app/src/main/kotlin/com/architects/pokearch/core/data/local/database/entities/converters/TypesHolder.kt
@@ -1,14 +1,0 @@
-package com.architects.pokearch.core.data.local.database.entities.converters
-
-data class TypesHolder (
-    val types: List<TypesEntity>
-)
-
-data class TypesEntity(
-    val slot: Int,
-    val type: TypeEntity
-)
-
-data class TypeEntity(
-    val name: String
-)
