+++ app/src/main/kotlin/com/architects/pokearch/core/framework/local/entities/converters/StatsHolder.kt
@@ -1,0 +1,12 @@
+package com.architects.pokearch.core.framework.local.entities.converters
+
+data class StatsHolder (val stats: List<StatsEntity>)
+
+data class StatsEntity(
+    val value: Int,
+    val stat: StatEntity
+)
+
+data class StatEntity(
+    val name: String
+)
