+++ app/src/main/kotlin/com/architects/pokearch/core/framework/local/entities/converters/TypesHolder.kt
@@ -1,0 +1,14 @@
+package com.architects.pokearch.core.framework.local.entities.converters
+
+data class TypesHolder (
+    val types: List<TypesEntity>
+)
+
+data class TypesEntity(
+    val slot: Int,
+    val type: TypeEntity
+)
+
+data class TypeEntity(
+    val name: String
+)
