+++ app/src/main/kotlin/com/architects/pokearch/core/framework/local/entities/converters/TypesHolderConverter.kt
@@ -1,0 +1,11 @@
+package com.architects.pokearch.core.framework.local.entities.converters
+
+import androidx.room.TypeConverter
+import com.google.gson.Gson
+
+class TypesHolderConverter {
+    @TypeConverter
+    fun fromType(types: TypesHolder): String = Gson().toJson(types)
+    @TypeConverter
+    fun toType(types: String): TypesHolder = Gson().fromJson(types, TypesHolder::class.java)
+}
