+++ app/src/main/kotlin/com/architects/pokearch/core/framework/mappers/EntityMapper.kt
@@ -1,0 +1,6 @@
+package com.architects.pokearch.core.framework.mappers
+
+interface EntityMapper<Network, Domain, Entity> {
+    fun asEntityFrom(network: Network): Entity
+    fun asDomainFrom(entity: Entity): Domain
+}
