+++ app/src/main/kotlin/com/architects/pokearch/core/framework/mappers/PokemonEntityMapper.kt
@@ -1,0 +1,27 @@
+package com.architects.pokearch.core.framework.mappers
+
+import com.architects.pokearch.core.framework.local.entities.PokemonEntity
+import com.architects.pokearch.core.framework.network.model.NetworkPokemon
+import com.architects.pokearch.core.domain.model.Pokemon
+
+object PokemonEntityMapper :
+    EntityMapper<List<NetworkPokemon>, List<Pokemon>, List<PokemonEntity>> {
+    override fun asEntityFrom(network: List<NetworkPokemon>): List<PokemonEntity> =
+        network.map(::mapToPokemonEntity)
+
+    private fun mapToPokemonEntity(networkPokemon: NetworkPokemon) =
+        PokemonEntity(
+            id = networkPokemon.getIndex(),
+            name = networkPokemon.name,
+            url = networkPokemon.url
+        )
+
+    override fun asDomainFrom(entity: List<PokemonEntity>): List<Pokemon> =
+        entity.map(::mapToPokemon)
+
+    private fun mapToPokemon(pokemonEntity: PokemonEntity) =
+        Pokemon(
+            name = pokemonEntity.name,
+            url = pokemonEntity.url
+        )
+}
