+++ app/src/main/kotlin/com/architects/pokearch/core/framework/network/model/NetworkPokemon.kt
@@ -1,0 +1,12 @@
+package com.architects.pokearch.core.framework.network.model
+
+import com.google.gson.annotations.SerializedName
+
+data class NetworkPokemon(
+    @SerializedName("name")
+    val name: String,
+    @SerializedName("url")
+    val url: String,
+) {
+    fun getIndex():Int = url.split("/".toRegex()).dropLast(1).last().toIntOrNull() ?: 0
+}
