+++ app/src/main/kotlin/com/architects/pokearch/core/framework/network/model/NetworkPokemons.kt
@@ -1,0 +1,14 @@
+package com.architects.pokearch.core.framework.network.model
+
+import com.google.gson.annotations.SerializedName
+
+data class NetworkPokemons(
+    @SerializedName("count")
+    val count: Int,
+    @SerializedName("next")
+    val next: String?,
+    @SerializedName("previous")
+    val previous: String?,
+    @SerializedName("results")
+    val results: List<NetworkPokemon>,
+)
